chore: reorganize tests in git package
