+++ packages/git/package.json
@@ -24,6 +24,8 @@
     "nock": "^10.0.6",
     "nyc": "^15.1.0",
     "oclif": "3.8.1",
+    "proxyquire": "^2.1.0",
+    "sinon": "^15.0.4",
     "ts-node": "^10.9.1",
     "tslib": "^1.9.3",
     "typescript": "4.8.4"
