+++ packages/git/test/unit/lib/git.unit.test.js
@@ -1,9 +1,9 @@
 'use strict'
-/* global describe it beforeEach afterEach */
+/* global beforeEach afterEach */
 
 let sinon = require('sinon')
-const { expect } = require('chai')
-let git = require('../../lib/git')({ httpGitHost: 'git.heroku.com', gitHost: 'heroku.com' })
+const {expect} = require('chai')
+let git = require('../../../lib/git')({httpGitHost: 'git.heroku.com', gitHost: 'heroku.com'})
 let cp = require('child_process')
 let EventEmitter = require('events')
 
@@ -17,14 +17,14 @@ describe('git', function () {
   it('runs exec', function () {
     mock.expects('execFile').withArgs('git', ['remote']).yieldsAsync(null, 'foo')
     return git.exec(['remote'])
-      .then((data) => {
-        expect(data.to.equal('foo')
+      .then(data => {
+        expect(data.to.equal('foo'))
         mock.verify()
       })
   })
 
   it('translates exec Errno::ENOENT to a friendlier error message', function () {
-    let err = new Error()
+    let err = new Error('err')
     err.code = 'ENOENT'
 
     mock.expects('execFile').withArgs('git', ['remote']).yieldsAsync(err, null)
@@ -42,18 +42,18 @@ describe('git', function () {
 
   it('runs spawn', function () {
     let emitter = new EventEmitter()
-    mock.expects('spawn').withExactArgs('git', ['remote'], { stdio: [0, 1, 2] }).returns(emitter)
+    mock.expects('spawn').withExactArgs('git', ['remote'], {stdio: [0, 1, 2]}).returns(emitter)
     process.nextTick(() => emitter.emit('close'))
     return git.spawn(['remote'])
       .then(() => mock.verify())
   })
 
   it('translates spawn Errno::ENOENT to a friendlier error message', function () {
-    let err = new Error()
+    let err = new Error('err')
     err.code = 'ENOENT'
 
     let emitter = new EventEmitter()
-    mock.expects('spawn').withExactArgs('git', ['remote'], { stdio: [0, 1, 2] }).returns(emitter)
+    mock.expects('spawn').withExactArgs('git', ['remote'], {stdio: [0, 1, 2]}).returns(emitter)
     process.nextTick(() => emitter.emit('error', err))
 
     return expect(git.spawn(['remote'])).to.be.rejectedWith('Git must be installed to use the Heroku CLI.  See instructions here: https://git-scm.com')
@@ -63,7 +63,7 @@ describe('git', function () {
     let err = new Error('Some other error message')
 
     let emitter = new EventEmitter()
-    mock.expects('spawn').withExactArgs('git', ['remote'], { stdio: [0, 1, 2] }).returns(emitter)
+    mock.expects('spawn').withExactArgs('git', ['remote'], {stdio: [0, 1, 2]}).returns(emitter)
     process.nextTick(() => emitter.emit('error', err))
 
     return expect(git.spawn(['remote'])).to.be.rejectedWith(err.message)
@@ -72,15 +72,15 @@ describe('git', function () {
   it('gets heroku git remote config', function () {
     mock.expects('execFile').withArgs('git', ['config', 'heroku.remote']).yieldsAsync(null, 'staging')
     return git.remoteFromGitConfig()
-      .then((remote) => expect(remote.to.equal('staging'))
+      .then(remote => expect(remote.to.equal('staging')))
       .then(() => mock.verify())
   })
 
   it('returns an http git url', function () {
-    expect(git.url('foo', false).to.equal('https://git.heroku.com/foo.git')
+    expect(git.url('foo', false).to.equal('https://git.heroku.com/foo.git'))
   })
 
   it('returns an ssh git url', function () {
-    expect(git.url('foo', true).to.equal('git@heroku.com:foo.git')
+    expect(git.url('foo', true).to.equal('git@heroku.com:foo.git'))
   })
 })
