+++ packages/git/test/unit/mock/git.js
@@ -2,15 +2,15 @@
 
 module.exports = {
   remoteFromGitConfig: () => Promise.resolve('heroku'),
-  url: (app) => `https://git.heroku.com/${app}.git`,
+  url: app => `https://git.heroku.com/${app}.git`,
   remoteUrl: () => Promise.resolve('https://git.heroku.com/myapp.git'),
   spawn: () => Promise.resolve(),
   exec: function (args) {
     switch (args.join(' ')) {
-      case 'remote':
-        return Promise.resolve('heroku')
-      default:
-        return Promise.resolve()
+    case 'remote':
+      return Promise.resolve('heroku')
+    default:
+      return Promise.resolve()
     }
-  }
+  },
 }
